requirements.txt
requests
django
djangorestframework
redis
celery
psycopg2
django-cors-headers
django-sendgrid-v5
